ShapeService and DataTypes to own file
